# internal/app/api/core/assets/doc/v0_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WireGuard Portal API - UI Endpoints",
        "title": "WireGuard Portal SPA-UI API",
        "contact": {
            "name": "WireGuard Portal Developers",
            "url": "https://github.com/fedor-git/wg-portal-2"
        },
        "version": "0.0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/user/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get all user records.",
                "operationId": "users_handleAllGet",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.User"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/new": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create the new user record.",
                "operationId": "users_handleCreatePost",
                "parameters": [
                    {
                        "description": "The user data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get a single user record.",
                "operationId": "users_handleSingleGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The user identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update the user record.",
                "operationId": "users_handleUpdatePut",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The user identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The user data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete the user record.",
                "operationId": "users_handleDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The user identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content if deletion was successful"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Error": {
            "type": "object",
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "Message": {
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "ApiEnabled": {
                    "type": "boolean"
                },
                "ApiToken": {
                    "type": "string"
                },
                "ApiTokenCreated": {
                    "type": "string"
                },
                "Department": {
                    "type": "string"
                },
                "Disabled": {
                    "description": "if this field is set, the user is disabled",
                    "type": "boolean"
                },
                "DisabledReason": {
                    "description": "the reason why the user has been disabled",
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "Firstname": {
                    "type": "string"
                },
                "Identifier": {
                    "type": "string"
                },
                "IsAdmin": {
                    "type": "boolean"
                },
                "Lastname": {
                    "type": "string"
                },
                "Locked": {
                    "description": "if this field is set, the user is locked",
                    "type": "boolean"
                },
                "LockedReason": {
                    "description": "the reason why the user has been locked",
                    "type": "string"
                },
                "Notes": {
                    "type": "string"
                },
                "Password": {
                    "type": "string"
                },
                "PeerCount": {
                    "type": "integer"
                },
                "Phone": {
                    "type": "string"
                },
                "ProviderName": {
                    "type": "string"
                },
                "Source": {
                    "type": "string"
                }
            }
        },
...
    }
}